--- packages/app-support/codemirror/.prettierignore
@@ -1,1 +1,0 @@
-../../../.prettierignore
