--- packages/app-support/codemirror/.prettierrc.cjs
@@ -1,1 +1,0 @@
-../../../.prettierrc.cjs
