--- packages/app-support/codemirror/CHANGELOG.md
@@ -1,11 +1,0 @@
-# @nullvoxpopuli/limber-codemirror
-
-## 0.0.1
-
-### Patch Changes
-
-- Updated dependencies []:
-  - @nullvoxpopuli/horizon-theme@0.0.1
-  - codemirror-lang-glimdown@0.0.0
-  - codemirror-lang-glimmer-js@0.0.0
-  - codemirror-lang-glimmer@0.0.0
