--- packages/app-support/codemirror/README.md
@@ -1,8 +1,0 @@
-# Custom CodeMirror build for Limber
-
-This package pre-builds CodeMirror so that
- - actual Limber code can be focused on Limber
- - keep package.json files concise
- - easy `await import(...)`(s)
-
-Output from `dist/` will be copied into the Limber app's `public` folder
