--- packages/app-support/codemirror/package.json
@@ -1,77 +1,0 @@
-{
-  "name": "@nullvoxpopuli/limber-codemirror",
-  "version": "0.0.1",
-  "description": "custom build of codemirror for use in Limber",
-  "type": "module",
-  "exports": {
-    "./preconfigured": "./dist/preconfigured.js",
-    "./broccoli-funnel": {
-      "require": "./broccoli-funnel.cjs"
-    }
-  },
-  "author": "NullVoxPopuli",
-  "license": "MIT",
-  "private": true,
-  "scripts": {
-    "build": "node build.cjs",
-    "dev": "pnpm build --watch",
-    "lint:types": "tsc --noEmit",
-    "lint:fix": "pnpm -w exec lint fix",
-    "lint": "pnpm -w exec lint",
-    "lint:prettier:fix": "pnpm -w exec lint prettier:fix",
-    "lint:prettier": "pnpm -w exec lint prettier"
-  },
-  "browser": {
-    "path": "path-browserify"
-  },
-  "dependencies": {
-    "@codemirror/autocomplete": "6.18.6",
-    "@codemirror/commands": "6.8.1",
-    "@codemirror/lang-html": "^6.4.9",
-    "@codemirror/lang-javascript": "6.2.4",
-    "@codemirror/lang-markdown": "6.3.3",
-    "@codemirror/lang-vue": "^0.1.3",
-    "@codemirror/lang-yaml": "^6.1.2",
-    "@codemirror/language": "^6.11.2",
-    "@codemirror/language-data": "^6.5.1",
-    "@codemirror/lint": "^6.8.4",
-    "@codemirror/search": "6.5.11",
-    "@codemirror/state": "6.5.2",
-    "@codemirror/view": "6.37.2",
-    "@lezer/common": "^1.2.3",
-    "@lezer/highlight": "^1.2.1",
-    "@lezer/html": "^1.3.10",
-    "@lezer/markdown": "^1.4.1",
-    "@nullvoxpopuli/horizon-theme": "workspace:*",
-    "@replit/codemirror-lang-svelte": "^6.0.0",
-    "assert": "npm:@nolyfill/assert@^1",
-    "broccoli-funnel": "^3.0.8",
-    "codemirror": "^6.0.2",
-    "codemirror-lang-glimdown": "workspace:*",
-    "codemirror-lang-glimmer": "workspace:*",
-    "codemirror-lang-glimmer-js": "workspace:*",
-    "codemirror-lang-mermaid": "^0.5.0",
-    "codemirror-languageserver": "^1.12.1",
-    "path-browserify": "^1.0.1"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.27.7",
-    "@nullvoxpopuli/eslint-configs": "^5.0.0",
-    "concurrently": "^9.1.2",
-    "esbuild": "0.25.5",
-    "esbuild-plugin-alias": "^0.2.1",
-    "eslint": "^9.29.0",
-    "prettier": "^3.6.2",
-    "prettier-plugin-ember-template-tag": "2.0.6",
-    "recursive-copy": "^2.0.14",
-    "typescript": "5.8.3"
-  },
-  "volta": {
-    "extends": "../../../package.json"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/NullVoxPopuli/limber.git",
-    "directory": "packages/app-support/codemirror"
-  }
-}
