+++ packages/repl-sdk/package.json
@@ -77,6 +77,30 @@
   },
   "dependencies": {
     "change-case": "^5.4.4",
+    "codemirror": "^6.0.2",
+    "@replit/codemirror-lang-svelte": "^6.0.0",
+    "@codemirror/autocomplete": "6.18.6",
+    "@codemirror/commands": "6.8.1",
+    "@codemirror/lang-html": "^6.4.9",
+    "@codemirror/lang-javascript": "6.2.4",
+    "@codemirror/lang-markdown": "6.3.3",
+    "@codemirror/lang-vue": "^0.1.3",
+    "@codemirror/lang-yaml": "^6.1.2",
+    "@codemirror/language": "^6.11.2",
+    "@codemirror/language-data": "^6.5.1",
+    "@codemirror/lint": "^6.8.4",
+    "@codemirror/search": "6.5.11",
+    "@codemirror/state": "6.5.2",
+    "@codemirror/view": "6.37.2",
+    "@lezer/common": "^1.2.3",
+    "@lezer/highlight": "^1.2.1",
+    "@lezer/html": "^1.3.10",
+    "@lezer/markdown": "^1.4.1",
+    "codemirror-lang-glimdown": "workspace:*",
+    "codemirror-lang-glimmer": "workspace:*",
+    "codemirror-lang-glimmer-js": "workspace:*",
+    "codemirror-lang-mermaid": "^0.5.0",
+    "codemirror-languageserver": "^1.12.1",
     "comlink": "^4.4.2",
     "es-module-shims": "^2.6.1",
     "mime": "^4.0.7",
