+++ packages/repl-sdk/src/compilers/markdown.js
@@ -23,6 +23,18 @@ export function filterOptions(options) {
  * @type {import('../types.ts').CompilerConfig}
  */
 export const md = {
+  codemirror: {
+    lang: async () => {
+      const [{ markdown }, { codeLanguages }] = await Promise.all([
+        import('@codemirror/lang-markdown'),
+        import('codemirror-lang-glimdown'),
+      ]);
+
+      return markdown({
+        codeLanguages,
+      });
+    },
+  },
   compiler: async (config, api) => {
     const { isLive, isPreview, needsLive, allowedFormats, isBelow, getFlavorFromMeta } =
       buildCodeFenceMetaUtils(api);
