+++ packages/repl-sdk/src/index.d.ts
@@ -1,4 +1,5 @@
 import type { CompilerConfig, Options } from './types.ts';
+import type { EditorView } from 'codemirror';
 
 export type { ErrorMessage, InfoMessage, Message, Options } from './types.ts';
 
@@ -18,4 +19,11 @@ export class Compiler {
   ): Promise<{ element: HTMLElement; destroy: () => void }>;
 
   optionsFor(format: string, flavor?: string): Omit<CompilerConfig, 'compiler'>;
+
+  createEditor(
+    element: HTMLElement,
+    text: string,
+    format: string,
+    handleUpdate: (updatedText: string) => void
+  ): Promise<{ view: EditorView; setText: (text: string, format: string) => Promise<void> }>;
 }
